Make sure to keep the size of your window less than half of the screen
This is essential as the structure of the website is such that if the window size is less than half of the screen, then for the purpose of finding the description, always a new tab would open up. Otherwise there will be 50% chances that the description loads on the same tab. To avoid this discrepancy, we must resize our tab window to whatever needed.